♻️refactor: CharacterLocker 관련 DTO 패키지로 따로 뺌 (#17)
