+++ src/main/java/kr/modernworld/modernworldv2/user/domain/characterlocker/CharacterLocker.java
@@ -1,0 +1,36 @@
+package kr.modernworld.modernworldv2.user.domain.characterlocker;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+public class CharacterLocker {
+
+  Long no;
+
+  Long userNo;
+
+  Long characterNo;
+
+  Boolean status;
+
+  public void equip() {
+    this.status = true;
+  }
+
+  public void unequip() {
+    this.status = false;
+  }
+
+  @Builder
+  private CharacterLocker(Long no, Long userNo, Long characterNo, Boolean status) {
+    this.no = no;
+    this.userNo = userNo;
+    this.characterNo = characterNo;
+    this.status = status;
+  }
+
+  public static CharacterLocker create(Long userNo, Long characterNo) {
+    return new CharacterLocker(null, userNo, characterNo, false);
+  }
+}
