+++ src/main/java/kr/modernworld/modernworldv2/user/presentation/characterlocker/dto/req/CharacterNoRequestDTO.java
@@ -1,4 +1,4 @@
-package kr.modernworld.modernworldv2.user.presentation.dto.req;
+package kr.modernworld.modernworldv2.user.presentation.characterlocker.dto.req;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
