+++ src/main/java/kr/modernworld/modernworldv2/user/presentation/characterlocker/dto/res/CharacterLockerResponseDTO.java
@@ -1,4 +1,4 @@
-package kr.modernworld.modernworldv2.user.presentation.dto.res;
+package kr.modernworld.modernworldv2.user.presentation.characterlocker.dto.res;
 
 public record CharacterLockerResponseDTO(
     Long no,
