+++ src/main/java/kr/modernworld/modernworldv2/user/presentation/characterlocker/dto/res/GetCharacterLockerResponseDTO.java
@@ -1,4 +1,4 @@
-package kr.modernworld.modernworldv2.user.presentation.dto.res;
+package kr.modernworld.modernworldv2.user.presentation.characterlocker.dto.res;
 
 public record GetCharacterLockerResponseDTO(
     Long no,
